Split Header/Class/Tag/Length to separate files
