+++ src/ber/length.rs
@@ -1,0 +1,11 @@
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub struct BerSizeError(pub(crate) ());
+
+/// Ber Object Length
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+pub enum BerSize {
+    /// Definite form (X.690 8.1.3.3)
+    Definite(usize),
+    /// Indefinite form (X.690 8.1.3.6)
+    Indefinite,
+}
